Add relay daemon mode to lobby_connect.
This (along with the existing custom_broadcasts support) allows for users on different read: _routed_ subnets to see and connect to each other's games.

To use this mode:

Build lobby_connect and run it with the "-d" argument.

Create the custom_broadcasts.txt file in the steam_settings folder next to steam_api.dll / steam_api64.dll and put the IP address of the host running lobby_connect in it.

Run games.

Note: This should even work with existing installations, _provided_ that at least one of them creates the custom_broadcasts.txt file as described above.
(This is because existing installations use broadcast packets to establish new connections, forwarding all of their knowledge to each other.
 Thus having at least one host on the current subnet configured in this new way turns it into a "proxy" for the other hosts.)

Caveat: This will _NOT_ work if NAT is involved. This is due to the lack of NAT traversal support in the emu.
(I.e. The lobby_connect daemon will get the info from the NAT'ed hosts and forward it to the others, but hosts outside the NAT won't be able to connect.)

Signed-off-by: redpolline <11156324-redpolline@users.noreply.gitlab.com>